+++ tools/scripts/vitest.config.ts
@@ -5,5 +5,9 @@ export default defineConfig({
     testTimeout: 60000,
     hookTimeout: 60000,
     include: ['test/**/*.test.ts'],
+    coverage: {
+      provider: 'v8',
+      exclude: ['**/test/**'],
+    },
   },
 });
